backend/requirements.txt
fastapi
uvicorn
openai
redis
sqlalchemy
python-multipart
pydantic
pydantic-settings
asyncio
websockets
langdetect
python-dotenv
